requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi @ file:///C:/Users/task_176606774961737/croot/argon2-cffi_1766067889655/work
argon2-cffi-bindings @ file:///C:/miniconda3/conda-bld/argon2-cffi-bindings_1757925012741/work
asttokens @ file:///C:/Users/task_177313475117182/croot/asttokens_1773135663340/work
async-lru @ file:///C:/miniconda3/conda-bld/async-lru_1761121600688/work
async-timeout==4.0.3
attrs==26.1.0
babel @ file:///C:/miniconda3/conda-bld/babel_1764159344759/work
bcrypt==5.0.0
beautifulsoup4 @ file:///C:/Users/task_177029095892244/croot/beautifulsoup4-split_1770291841733/work
bleach @ file:///C:/miniconda3/conda-bld/bleach_1764153707707/work
blinker==1.9.0
brotlicffi @ file:///C:/miniconda3/conda-bld/brotlicffi_1764961372591/work
build==1.4.4
cachetools==7.0.6
certifi==2026.4.22
cffi @ file:///C:/miniconda3/conda-bld/cffi_1761832794697/work
charset-normalizer==3.4.7
chromadb==1.5.8
click==8.3.3
colorama==0.4.6
coloredlogs==15.0.1
comm @ file:///C:/miniconda3/conda-bld/comm_1763119233835/work
dataclasses-json==0.6.7
debugpy @ file:///C:/miniconda3/conda-bld/debugpy_1762421657095/work
decorator @ file:///C:/miniconda3/conda-bld/decorator_1757341256088/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.1
executing @ file:///C:/miniconda3/conda-bld/executing_1757061252959/work
fastapi==0.136.1
fastjsonschema @ file:///C:/miniconda3/conda-bld/python-fastjsonschema_1763718726757/work
filelock==3.29.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.47
googleapis-common-protos==1.74.0
greenlet==3.4.0
grpcio==1.80.0
h11==0.16.0
html5lib @ file:///Users/ktietz/demo/mc3/conda-bld/html5lib_1629144453894/work
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.13
importlib_metadata==8.7.1
importlib_resources==7.1.0
ipykernel @ file:///C:/Users/task_177261243501734/croot/ipykernel_1772612536641/work
ipython @ file:///C:/b/abs_8eyhzleyrk/croot/ipython_1734548134403/work
jedi @ file:///C:/b/abs_3a2kbnlclc/croot/jedi_1733987412687/work
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
json5 @ file:///C:/miniconda3/conda-bld/json5_1763718570061/work
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events @ file:///C:/miniconda3/conda-bld/jupyter_events_1765813622940/work
jupyter-lsp @ file:///C:/Users/task_176977265234360/croot/jupyter-lsp_1769772755702/work
jupyter_client @ file:///C:/Users/task_176839864521852/croot/jupyter_client_1768399093953/work
jupyter_core @ file:///C:/miniconda3/conda-bld/jupyter_core_1764588347534/work
jupyter_server @ file:///C:/Users/task_176961323688807/croot/jupyter_server_1769613449402/work
jupyter_server_terminals @ file:///C:/Users/task_177028940194448/croot/jupyter_server_terminals_1770289524213/work
jupyterlab @ file:///C:/Users/task_177512637369589/croot/jupyterlab_1775126465557/work
jupyterlab_pygments @ file:///C:/b/abs_d5alfet8m6/croot/jupyterlab_pygments_1741124274578/work
jupyterlab_server @ file:///C:/Users/task_176962937278440/croot/jupyterlab_server_1769629917911/work
kubernetes==35.0.0
langchain-chroma==1.1.0
langchain-classic==1.0.4
langchain-community==0.4.1
langchain-core==1.3.2
langchain-protocol==0.0.12
langchain-text-splitters==1.1.2
langgraph==1.1.9
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.11
langgraph-sdk==0.3.13
langsmith==0.7.36
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline @ file:///C:/miniconda3/conda-bld/matplotlib-inline_1762779265381/work
mdurl==0.1.2
mistune @ file:///C:/b/abs_77yql3poyz/croot/mistune_1741124004410/work
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.20.0
nbclient @ file:///C:/Users/task_177157717480981/croot/nbclient_1771577305053/work
nbconvert @ file:///C:/Users/task_177212126378136/croot/nbconvert-meta_1772121327883/work
nbformat @ file:///C:/b/abs_c2jkw46etm/croot/nbformat_1728050303821/work
neo4j==5.28.3
nest-asyncio @ file:///C:/b/abs_65d6lblmoi/croot/nest-asyncio_1708532721305/work
networkx==3.4.2
notebook @ file:///C:/Users/task_177513921688324/croot/notebook_1775139388341/work
notebook_shim @ file:///C:/Users/task_176962932903714/croot/notebook-shim_1769629908104/work
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.23.2
openai==1.59.9
opentelemetry-api==1.41.1
opentelemetry-exporter-otlp-proto-common==1.41.1
opentelemetry-exporter-otlp-proto-grpc==1.41.1
opentelemetry-proto==1.41.1
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
orjson==3.11.8
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
pandas==2.3.3
pandocfilters @ file:///C:/miniconda3/conda-bld/pandocfilters_1756977625119/work
parso @ file:///C:/miniconda3/conda-bld/parso_1762781925562/work
pillow==12.2.0
platformdirs @ file:///C:/Users/task_177365233215608/croot/platformdirs_1773652995602/work
prometheus_client @ file:///C:/Users/task_177273314439581/croot/prometheus_client_1772733422509/work
prompt_toolkit @ file:///C:/miniconda3/conda-bld/prompt-toolkit_1761745021230/work
propcache==0.4.1
protobuf==6.33.6
psutil @ file:///C:/miniconda3/conda-bld/psutil_1761896539676/work
pure_eval @ file:///C:/miniconda3/conda-bld/pure_eval_1757067068391/work
pyarrow==24.0.0
pybase64==1.4.3
pycparser @ file:///C:/Users/task_177495971250430/croot/pycparser_1774959822609/work
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
pydeck==0.9.2
Pygments==2.20.0
pypdf==4.3.1
PyPika==0.51.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks @ file:///C:/miniconda3/conda-bld/pysocks_1761753030270/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger @ file:///C:/miniconda3/conda-bld/python-json-logger_1764590249459/work
pytz==2026.1.post1
pywin32 @ file:///C:/b/abs_6c9w_vp9zi/croot/pywin32_1754670534373/work
pywinpty @ file:///C:/Users/task_177521346875498/croot/pywinpty_1775213569613/work/pywinpty-src
PyYAML==6.0.3
pyzmq @ file:///C:/miniconda3/conda-bld/pyzmq_1762375610449/work
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator @ file:///C:/b/abs_ddfmseb_vm/croot/rfc3339-validator_1683077054906/work
rfc3986-validator @ file:///C:/b/abs_6e9azihr8o/croot/rfc3986-validator_1683059049737/work
rich==15.0.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.15.3
Send2Trash @ file:///C:/Users/task_176817065969485/croot/send2trash_1768170763283/work
sentence-transformers==3.4.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve @ file:///C:/b/abs_bbsvy9t4pl/croot/soupsieve_1696347611357/work
SQLAlchemy==2.0.49
stack_data @ file:///C:/miniconda3/conda-bld/stack_data_1757067051465/work
starlette==1.0.0
streamlit==1.56.0
sympy==1.14.0
tenacity==9.1.4
terminado @ file:///C:/Users/task_176959131752406/croot/terminado_1769591781114/work
threadpoolctl==3.6.0
tinycss2 @ file:///C:/b/abs_df38owi5ma/croot/tinycss2_1738337725183/work
tokenizers==0.21.4
toml==0.10.2
tomli==2.4.1
torch==2.11.0
tornado @ file:///C:/Users/task_177430356365515/croot/tornado_1774303657186/work
tqdm==4.67.3
traitlets @ file:///C:/b/abs_bfsnoxl4pq/croot/traitlets_1718227069245/work
transformers==4.48.3
typer==0.25.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.46.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth @ file:///C:/Users/task_176777900206875/croot/wcwidth_1767779217517/work
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
win_inet_pton @ file:///C:/miniconda3/conda-bld/win_inet_pton_1761746276406/work
xxhash==3.7.0
yarl==1.23.0
zipp==3.23.1
zstandard==0.25.0
